Complete restructuring app into pages and components
