+++ public/images/idyeahsLogoWhite.png
