+++ public/index.html
@@ -2,15 +2,14 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
-
     <meta
       name="viewport"
       content="width=device-width,
-      initial-scale=1"
+    initial-scale=1"
     />
     <meta name="theme-color" content="#000000" />
-
+    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico" />
+    <link rel="stylesheet" href="%PUBLIC_URL%/style.css" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <link
       rel="stylesheet"
@@ -20,11 +19,11 @@
       rel="stylesheet"
       href="https://fonts.googleapis.com/css?family=Rubik:300,400,500&display=swap"
     />
-
     <title>IDYEAHS</title>
   </head>
+
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
+    <noscript>You'll need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
   </body>
 </html>
