+++ src/components/Comment.js
@@ -4,8 +4,8 @@ import OutsideClickHandler from 'react-outside-click-handler';
 import moment from 'moment';
 import { FontAwesomeIcon as Icon } from '@fortawesome/react-fontawesome';
 import { faPaperPlane } from '@fortawesome/free-solid-svg-icons';
-import ContextMenuComment from './ContextMenuComment';
-import DeleteCommentModal from './DeleteCommentModal';
+import ContextMenuComment from '../components/ContextMenuComment';
+import DeleteCommentModal from '../components/DeleteCommentModal';
 import Input from '../components/form/Input';
 import { CommentAvatar } from '../components/Avatar';
 moment.locale('de');
