+++ src/data/ActiveUser.js
@@ -4,7 +4,7 @@ const aUser = {
   lastname: 'Clarke',
   email: 'leo@idyeahs.com',
   avatar: {
-    src: '/img/leo.jpg',
+    src: '/images/leo.jpg',
     alt: 'Leo Avatar Image',
   },
   pwd: '123',
