+++ src/data/Users.js
@@ -7,7 +7,7 @@ const usersList = [
     role: 'Helpdesk Technician',
     department: 'Helpdesk',
     avatar: {
-      src: '/img/leo.jpg',
+      src: '/images/leo.jpg',
       alt: 'Leo Avatar Image',
     },
     pwd: '123',
@@ -20,7 +20,7 @@ const usersList = [
     role: 'Pilot',
     department: 'outpost b19',
     avatar: {
-      src: '/img/macready.jpg',
+      src: '/images/macready.jpg',
       alt: 'MacReady Avatar Image',
     },
     pwd: '123',
@@ -33,7 +33,7 @@ const usersList = [
     role: 'software developer',
     department: 'python',
     avatar: {
-      src: '/img/defaultAvatar.png',
+      src: '/images/defaultAvatar.png',
       alt: 'default Avatar',
     },
     pwd: '123',
