+++ src/pages/IdeaDetailsView.js
