+++ src/pages/IdeasFeed.js
@@ -1,72 +1,50 @@
 import React from 'react';
-import styled from 'styled-components';
-import { NavLink } from 'react-router-dom';
-import { library } from '@fortawesome/fontawesome-svg-core';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTimes } from '@fortawesome/free-solid-svg-icons';
-import IdeaPost from '../app/IdeaPost';
-
-library.add(faTimes);
-
-const StyledSection = styled.div`
-  padding: 10px;
-  overflow: scroll;
-`;
-
-const FilterTitle = styled.div`
-  color: rebeccapurple;
-`;
-
-const StyledTag = styled.span`
-  padding: 3px 6px;
-  margin-right: 5px;
-  background: #7fb7be;
-  color: white;
-  border-radius: 2px;
-  font-size: 0.7em;
-`;
-
-const IdeaDetailsLink = styled(NavLink)`
-  color: white;
-  text-decoration: none;
-  &.link,
-  .visited,
-  .hover,
-  .active {
-  }
-`;
+import Header from '../components/Header';
+import IdeaCard from '../components/IdeaCard';
+import { AddButton } from '../components/FAB';
+import { MainGrid, FeedGridIdeas } from '../components/Grids.elements';
 
 export default function IdeasFeed({
+  heading,
   posts,
-  tagFilter,
-  onTagClick,
-  resetFilter,
+  activeUser,
+  onContextClick,
+  onIdeaDelete,
+  onStarAdd,
+  onStarRemove,
+  history,
+  showStarred,
 }) {
+  function getFilteredPosts() {
+    return posts.filter(post => {
+      return post.stars.find(
+        star => star.author.username === activeUser.username
+      );
+    });
+  }
+  const displayPosts = showStarred ? getFilteredPosts() : posts;
   return (
-    <StyledSection>
-      {tagFilter && (
-        <FilterTitle>
-          {tagFilter.map(tagName => (
-            <StyledTag key={tagName.id}>{tagName}</StyledTag>
-          ))}
-          <FontAwesomeIcon icon="times" onClick={resetFilter} />
-        </FilterTitle>
-      )}
-      {posts
-        .filter(post => post.tags.toString().includes(tagFilter))
-        .map(post => (
-          <IdeaDetailsLink to={`/ideas/details/${post.id}`} key={post.id}>
-            <IdeaPost
-              key={post.id}
-              title={post.title}
-              text={post.text}
-              tags={post.tags}
-              timestamp={post.timestamp}
-              username={post.username}
-              onTagClick={onTagClick}
-            />
-          </IdeaDetailsLink>
+    <MainGrid>
+      <Header
+        heading={heading}
+        history={history}
+        activeUser={activeUser}
+        fab={!showStarred && <AddButton />}
+      />
+      <FeedGridIdeas>
+        {displayPosts.map(post => (
+          <IdeaCard
+            key={post.id}
+            post={post}
+            activeUser={activeUser}
+            onContextClick={onContextClick}
+            onStarAdd={onStarAdd}
+            onStarRemove={onStarRemove}
+            onIdeaDelete={onIdeaDelete}
+            history={history}
+          />
         ))}
-    </StyledSection>
+      </FeedGridIdeas>
+    </MainGrid>
   );
 }
